Add caching middleware
Calls to functions wrapped with `wrap-cache` are cached to files.
By default these are stored in .karma-cache (configurable with environment variables).
Running `lein clean` will delete the cache directory.

`wrap-cache` may be used as clj-http middleware by passing it to `with-middleware`.
It must be inserted at the end of the middleware chain, after the response has been parsed.

    (http/with-middleware (conj http/default-middleware wrap-cache) ...)