Added `augmentation` layer.
This layer works on the aggregated data in order to add new info.
Some of them of downloaded from GitHub and some of them are calculated
from data already in the report.

One example is the data about the languages used. GitHub returns the
amount of bytes written for every language. We need percentages, so
the augmentation layer compute those values needed by the reports.